Refactor Subscription service and add linkify dependency
Refactored subscription logic to isolate DB write and mail sending into their functions for better code organization and error handling. This also includes modification of error handling strategy to enhance traceability in the system. The new function `handle_email_error` is created for this purpose. The linkify lib was added to `Cargo.toml` in order to validate confirmation link in tests. A new `full_url` method was added to `ApplicationSettings` to compute the full application URL from the configuration. Application scheme was also added to the test configuration.

These changes improve the organization of the code, making it easier to trace, test and maintain. They also ensure the URLs used in emails are valid.