Add migration operation option to API server
This commit adds a migration operation to the API server's main function. It checks for a 'migrate' argument in the command line and runs database migration if present. Besides, several previously deleted and then added .json query files are re-introduced. Lastly, a new 'migrate.rs' file is created to handle application migrations.