Implement feature toggle for local and shuttle environments
This commit introduces a feature toggle to separate code execution for local and shuttle deployments. Using cargo features, conditional sections of code have been created in both the backend and server main files. Moreover, the shuttle libraries are made optional dependencies in Cargo.toml, which are only compiled when the `shuttle` feature flag is on. Conversely, specific code for local deployment is only activated with the `local` feature flag. To support these changes, appropriate commands have been added to the Makefile.