Refactor frontend to use Sycamore instead of Dioxus for building
The frontend was transitioned from Dioxus to Sycamore for the building process. This resulted in the deletion of Dioxus-specific files and addition of new configs for Sycamore integration. Several updates were made to accommodate this shift including alterations to the Makefile, .gitignore file, Cargo.toml, and Cargo.lock. Also, the frontend page content was updated for Sycamore components and routing.