Upgrade Dockerfile and add run_release to Makefile
Enhancements have been done in the Dockerfile to include installing 'trunk', adding the 'wasm32-unknown-unknown' target, and building frontend in release mode. Further, a new command `run_release` has been added to the Makefile to allow building the frontend in release mode. This streamlines the process for launching the production version of the app.