Implement user confirmation and navigation enhancements
This commit adds a confirmation mechanism for users. A new 'confirm' module has been added to the front-end and backend services for this purpose. It also includes updates to the navigation system for handling user redirections and introduces a new navbar for authenticated users. Adjustments are made to the build-front-end command in the Makefile, allowing to turn off filehash. Test cases for user confirmations in the backend services were also updated.